YT-28280: fix integer overflow commit_hash:cdc2688c3b299db04889e6740e294aff2c8df6f3

+++ yt/yt/server/node/exec_node/layer_location.cpp
@@ -215,8 +215,8 @@ TFuture<TVolumeMeta> TLayerLocation::CreateOverlayVolume(
     TEventTimerGuard volumeCreateTimeGuard,
     int userId,
     const std::optional<std::string>& placePath,
-    std::optional<int> diskSpaceLimit,
-    std::optional<int> inodeLimit,
+    std::optional<i64> diskSpaceLimit,
+    std::optional<i64> inodeLimit,
     const std::vector<TOverlayData>& overlayDataArray,
     bool placeInUserSlot)
 {
@@ -1115,8 +1115,8 @@ TVolumeMeta TLayerLocation::DoCreateOverlayVolume(
     TEventTimerGuard volumeCreateTimeGuard,
     int userId,
     const std::optional<std::string>& placePath,
-    std::optional<int> diskSpaceLimit,
-    std::optional<int> inodeLimit,
+    std::optional<i64> diskSpaceLimit,
+    std::optional<i64> inodeLimit,
     const std::vector<TOverlayData>& overlayDataArray,
     bool placeInUserSlot)
 {
