+++ yt/yt/server/node/exec_node/layer_location.h
@@ -63,8 +63,8 @@ public:
         TEventTimerGuard volumeCreateTimeGuard,
         int userId,
         const std::optional<std::string>& placePath,
-        std::optional<int> diskSpaceLimit,
-        std::optional<int> inodeLimit,
+        std::optional<i64> diskSpaceLimit,
+        std::optional<i64> inodeLimit,
         const std::vector<TOverlayData>& overlayDataArray,
         bool placeInUserSlot);
 
@@ -241,8 +241,8 @@ private:
         TEventTimerGuard volumeCreateTimeGuard,
         int userId,
         const std::optional<std::string>& placePath,
-        std::optional<int> diskSpaceLimit,
-        std::optional<int> inodeLimit,
+        std::optional<i64> diskSpaceLimit,
+        std::optional<i64> inodeLimit,
         const std::vector<TOverlayData>& overlayDataArray,
         bool placeInUserSlot);
